requirements.txt
fastapi
uvicorn
python-multipart
dicom2nifti
nibabel
pyrobex
Pillow
numpy
